requirements.txt
asgiref==3.8.1
colorama==0.4.6
coverage==7.6.12
Django==5.1.6
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-seed==0.3.1
djangorestframework==3.15.2
Faker==36.1.1
iniconfig==2.0.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.10.0
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
sqlparse==0.5.3
toposort==1.10
typing_extensions==4.12.2
tzdata==2025.1
